# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: H&CN Studio API - Product detail page creation system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /video-tasks:
    get:
      operationId: listVideoTasks
      tags: [video-tasks]
      summary: List all video tasks (newest first)
      responses:
        "200":
          description: List of video tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoTask"
    post:
      operationId: createVideoTask
      tags: [video-tasks]
      summary: Create a new video task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoTaskBody"
      responses:
        "201":
          description: Created video task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTask"
  /video-tasks/{id}:
    get:
      operationId: getVideoTask
      tags: [video-tasks]
      summary: Get video task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTask"
        "404":
          description: Not found
    patch:
      operationId: updateVideoTask
      tags: [video-tasks]
      summary: Update a video task name
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVideoTaskBody"
      responses:
        "200":
          description: Updated video task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTask"
        "404":
          description: Not found
    delete:
      operationId: deleteVideoTask
      tags: [video-tasks]
      summary: Delete a video task (nullifies videoTaskId on associated videos)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    VideoTask:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    CreateVideoTaskBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UpdateVideoTaskBody:
      type: object
      properties:
        name:
          type: string
...