# lib/api-zod/src/generated/types/updateVideoTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * H&CN Studio API - Product detail page creation system
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateVideoTaskBody {
  name?: string;
}